First MonaServer2 version with MonaServer project and LUA engine which builds on WIN32 (linux build is coming)

+++ MonaTiny/sources/main.cpp
@@ -19,7 +19,7 @@ struct ServerApp : ServerApplication  {
 	int main(TerminateSignal& terminateSignal) {
 
 		// starts the server
-		MonaTiny server(terminateSignal, getNumber<UInt32>("cores"));
+		MonaTiny server(terminateSignal, getNumber<UInt16>("cores"));
 
 		server.start(*this);
